removed unused interface competition, moved community
