+++ source/battle-board/battle-board/app/search/page.tsx
@@ -5,7 +5,7 @@ import CompetitonSearchItem from "@/components/search/competition-search-item";
 import CommunitySearchItem from "@/components/search/community-search-item";
 
 import CompetitionData from "@/models/interfaces/CompetitionData";
-import Community from "@/models/community";
+import Community from "@/models/interfaces/community";
 
 import { useEffect, useState } from "react";
 import ButtonData from "@/models/component-props/button-data";
