--- source/battle-board/battle-board/models/competition.ts
@@ -1,8 +1,0 @@
-export default interface Competition {
-  id: number;
-  title: string;
-  game: string;
-  participants: number;
-  competition_type: string;
-  start_date: string;
-}
