+++ source/battle-board/battle-board/models/interfaces/community.ts
